Use 64 integers and serialize w/o the loop

+++ include/move_query.hpp
@@ -7,12 +7,12 @@ class MoveQuery {
         MoveQuery(std::string q) : query_string(q) {}
 
         std::string query() { return query_string; }
-        uint32_t length() { return query_string.length(); }
-        void add_ms(uint32_t len) { ms_lens.push_back(len); }
+        uint64_t length() { return query_string.length(); }
+        void add_ms(uint64_t len) { ms_lens.push_back(len); }
 
         friend std::ostream& operator<<(std::ostream& output, const MoveQuery& dt) {
             output << "The matching statistics are:\n";
-            for (int32_t i = dt.ms_lens.size() - 1; i >= 0; i--) {
+            for (int64_t i = dt.ms_lens.size() - 1; i >= 0; i--) {
                 auto ms = dt.ms_lens[i];
                 output << ms << " ";
             }
@@ -20,7 +20,7 @@ class MoveQuery {
         }
     private:
         std::string query_string;
-        std::vector<uint32_t> ms_lens;
+        std::vector<uint64_t> ms_lens;
 };
 
 #endif
